Major refactoring. As part of this, the commented out and incomplete exporters (bezier, armature) were removed
