+++ io_scene_godot/converters/__init__.py
@@ -1,0 +1,11 @@
+"""
+This file provides the conversion for a single blend object into one or
+more godot nodes. Most of the functions in here take the file, a node, and
+the global path to the node.
+
+one-function exporters are stored in simple_nodes. Others (such as meshes)
+are stored in individual files
+"""
+
+from .simple_nodes import *
+from .mesh import export_mesh_node
